fix(test): run cabal update prior to GitHub Action tests
This is due to https://github.com/haskell/cabal/pull/8944

Previously, we were getting only a warning. It was fine as we are using
Nix to provision the packages.

Now (as of cabal-install v3.12.1.0), we are getting an error. To avoid
this, we need to run `cabal update` prior to running the tests.

Note that there are ongoing discussions on how to handle this going
forward.